+++ nelumbo-server/src/test/java/org/modelingvalue/nelumbo/server/InferenceDeadlineTest.java
@@ -14,7 +14,7 @@
 //     Victor Lap                                                                                                      ~
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-package org.modelingvalue.nelumbo.website;
+package org.modelingvalue.nelumbo.server;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;